refactor(infra): T00 establish architecture convention and infrastructure layer
- Add CLAUDE.md documenting feature-first layering, dependency-direction
  rules, the infrastructure/ layer, and code-generation workflow.
- Introduce lib/infrastructure/ and relocate domain-agnostic foundation
  providers from lib/providers/:
  firebase_providers, firestore_providers, core_providers, pref_provider,
  firestore_error_notifier (with their generated .g.dart pairs).
- Update all imports (providers/ -> infrastructure/) across features,
  pages, components, repositories and remaining provider files; fix import
  ordering.

No behavior change. flutter analyze unchanged (337 info-level, pre-existing);
build_runner regenerates with no diff.
