+++ lib/features/auth/presentation/email_link_auth_notifier.dart
@@ -2,7 +2,7 @@ import "package:freezed_annotation/freezed_annotation.dart";
 import "package:riverpod_annotation/riverpod_annotation.dart";
 
 import "../../../core/pref_key.dart";
-import "../../../providers/firebase_providers.dart";
+import "../../../infrastructure/firebase_providers.dart";
 import "../../../providers/link_events_provider.dart";
 import "../../../utils/app_links.dart";
 import "../../../utils/notifier_state_guard.dart";
