+++ lib/features/auth/repositories/auth_repository.dart
@@ -8,8 +8,8 @@ import "package:riverpod_annotation/riverpod_annotation.dart";
 import "package:sign_in_with_apple/sign_in_with_apple.dart";
 
 import "../../../apple_web_auth_options.dart";
-import "../../../providers/core_providers.dart";
-import "../../../providers/firebase_providers.dart";
+import "../../../infrastructure/core_providers.dart";
+import "../../../infrastructure/firebase_providers.dart";
 import "../../../utils/app_links.dart";
 import "../models/auth_exception.dart";
 import "apple_sign_in_android.dart";
