+++ lib/features/google_tasks/repositories/tasks_auth_notifier.dart
@@ -6,8 +6,8 @@ import "package:googleapis/tasks/v1.dart";
 import "package:googleapis_auth/googleapis_auth.dart";
 import "package:riverpod_annotation/riverpod_annotation.dart";
 
-import "../../../providers/core_providers.dart";
-import "../../../providers/firebase_providers.dart";
+import "../../../infrastructure/core_providers.dart";
+import "../../../infrastructure/firebase_providers.dart";
 import "../models/tasks_auth_exception.dart";
 
 part "tasks_auth_notifier.g.dart";
