+++ lib/infrastructure/firestore_error_notifier.dart
