+++ lib/main.dart
@@ -26,6 +26,10 @@ import "features/auth/presentation/auth_messages.dart";
 import "features/auth/presentation/email_link_auth_notifier.dart";
 import "features/google_tasks/models/tasks_operation_exception.dart";
 import "features/submission/presentation/submission_save_state_notifier.dart";
+import "infrastructure/core_providers.dart";
+import "infrastructure/firebase_providers.dart";
+import "infrastructure/firestore_error_notifier.dart";
+import "infrastructure/pref_provider.dart";
 import "link_handler/link_handler.dart";
 import "pages/done_submissions_page.dart";
 import "pages/email_registration_page.dart";
@@ -47,11 +51,7 @@ import "pages/submission_edit_page.dart";
 import "pages/timetable_edit_page.dart";
 import "pages/timetable_table_view_page.dart";
 import "pages/welcome_page.dart";
-import "providers/core_providers.dart";
-import "providers/firebase_providers.dart";
-import "providers/firestore_error_notifier.dart";
 import "providers/link_events_provider.dart";
-import "providers/pref_provider.dart";
 import "utils/distinguish.dart";
 import "utils/ui.dart";
 import "utils/utils.dart";
