+++ lib/pages/home_page.dart
@@ -14,11 +14,11 @@ import "../core/pref_key.dart";
 import "../events.dart";
 import "../fade_through_page_route.dart";
 import "../features/auth/use_cases/sign_out_use_case.dart";
+import "../infrastructure/core_providers.dart";
+import "../infrastructure/firebase_providers.dart";
 import "../isar_db/isar_digestive.dart";
-import "../providers/core_providers.dart";
 import "../providers/data_sync_service.dart";
 import "../providers/digestive_providers.dart";
-import "../providers/firebase_providers.dart";
 import "../src/pigeons.g.dart";
 import "../ui_components/hidable_progress_indicator.dart";
 import "../utils/ad_unit_ids.dart";
