+++ lib/pages/settings/functions.dart
@@ -9,8 +9,8 @@ import "../../browser.dart";
 import "../../components/dropdown_time_picker_bottom_sheet.dart";
 import "../../features/auth/use_cases/common.dart";
 import "../../features/auth/use_cases/sign_out_use_case.dart";
-import "../../providers/firebase_providers.dart";
-import "../../providers/firestore_providers.dart";
+import "../../infrastructure/firebase_providers.dart";
+import "../../infrastructure/firestore_providers.dart";
 import "../../src/pigeons.g.dart";
 import "../../ui_components/settings_ui.dart";
 import "../../utils/ui.dart";
