+++ lib/pages/submission_detail_page.dart
@@ -8,10 +8,10 @@ import "package:url_launcher/url_launcher_string.dart";
 import "../components/digestive_detail_card.dart";
 import "../components/digestive_edit_bottom_sheet.dart";
 import "../components/submissions/formatted_date_remaining.dart";
+import "../infrastructure/firebase_providers.dart";
 import "../isar_db/isar_digestive.dart";
 import "../isar_db/isar_submission.dart";
 import "../providers/digestive_providers.dart";
-import "../providers/firebase_providers.dart";
 import "../providers/submission_providers.dart";
 import "../utils/ad_unit_ids.dart";
 import "../utils/ui.dart";
