+++ lib/pages/welcome_page.dart
@@ -8,8 +8,8 @@ import "../browser.dart";
 import "../core/pref_key.dart";
 import "../features/auth/repositories/auth_repository.dart";
 import "../features/auth/use_cases/common.dart";
-import "../providers/firebase_providers.dart";
-import "../providers/firestore_providers.dart";
+import "../infrastructure/firebase_providers.dart";
+import "../infrastructure/firestore_providers.dart";
 import "../utils/ui.dart";
 import "home_page.dart";
 import "sign_in_page.dart";
