+++ lib/providers/data_sync_service.dart
@@ -6,6 +6,9 @@ import "package:cloud_firestore/cloud_firestore.dart";
 import "package:riverpod_annotation/riverpod_annotation.dart";
 
 import "../core/pref_key.dart";
+import "../infrastructure/core_providers.dart";
+import "../infrastructure/firebase_providers.dart";
+import "../infrastructure/firestore_providers.dart";
 import "../isar_db/isar_digestive.dart";
 import "../isar_db/isar_submission.dart";
 import "../isar_db/isar_timetable.dart";
@@ -14,9 +17,6 @@ import "../isar_db/isar_timetable_table.dart";
 import "../user_config.dart";
 import "../utils/batch_operation.dart";
 import "../utils/notifier_state_guard.dart";
-import "core_providers.dart";
-import "firebase_providers.dart";
-import "firestore_providers.dart";
 
 part "data_sync_service.g.dart";
