+++ lib/providers/digestive_providers.dart
@@ -1,14 +1,14 @@
 import "package:isar_community/isar.dart";
 import "package:riverpod_annotation/riverpod_annotation.dart";
 
+import "../infrastructure/core_providers.dart";
+import "../infrastructure/firebase_providers.dart";
+import "../infrastructure/firestore_error_notifier.dart";
+import "../infrastructure/firestore_providers.dart";
 import "../isar_db/isar_digestive.dart";
 import "../isar_db/isar_submission.dart";
 import "../pages/home_tabs/tab_digestive_list.dart";
 import "../repositories/digestive_repository.dart";
-import "core_providers.dart";
-import "firebase_providers.dart";
-import "firestore_error_notifier.dart";
-import "firestore_providers.dart";
 
 part "digestive_providers.g.dart";
