+++ lib/providers/timetable_providers.dart
@@ -2,16 +2,16 @@ import "package:isar_community/isar.dart";
 import "package:riverpod_annotation/riverpod_annotation.dart";
 
 import "../core/pref_key.dart";
+import "../infrastructure/core_providers.dart";
+import "../infrastructure/firebase_providers.dart";
+import "../infrastructure/firestore_error_notifier.dart";
+import "../infrastructure/firestore_providers.dart";
 import "../isar_db/isar_timetable.dart";
 import "../isar_db/isar_timetable_class_time.dart";
 import "../isar_db/isar_timetable_table.dart";
 import "../repositories/timetable_class_time_repository.dart";
 import "../repositories/timetable_repository.dart";
 import "../repositories/timetable_table_repository.dart";
-import "core_providers.dart";
-import "firebase_providers.dart";
-import "firestore_error_notifier.dart";
-import "firestore_providers.dart";
 
 part "timetable_providers.g.dart";
